[C#] Option to create barrier at checkpoint version switch, support rollback (#813)
* Whether we use a barrier to ensure that threads are not in two different checkpoint versions at the same time. In PREPARE phase, refresh will wait for threads to get to the next version.

* Add callback to recognize version shift during checkpointing

* Add API to FasterLog to initialize empty log with given address.

* Support basic recovery rollback, just call Recover on an existing store, and it will undo any uncheckpointed data. Make sure store is quiesced before attempting to call Recover on it.

* Ensure proper version switch

* fix grow loop

* Make grow increment version as well. Address comments.

Basic idea:

Threads in PREPARE_* phase will proceed as normal if NumActiveLockingSessions > 0
Else they will make the state machine move forward to IN_PROG_*

+++ cs/src/core/Index/Synchronization/VersionChangeStateMachine.cs
@@ -147,9 +147,8 @@ namespace FASTER.core
                     break;
                 case Phase.PREPARE:
                     nextState.Phase = Phase.IN_PROGRESS;
-                    // TODO: Move to long for system state as well. 
                     SetToVersion(targetVersion == -1 ? start.Version + 1 : targetVersion);
-                    nextState.Version = (int)ToVersion();
+                    nextState.Version = ToVersion();
                     break;
                 case Phase.IN_PROGRESS:
                     nextState.Phase = Phase.REST;
